+++ js/modules/constant.js
@@ -1,0 +1,53 @@
+const photos = [];
+const photosUrl = [];
+const photosId = [];
+const commentId = [];
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const NAMES = [
+  'Миша',
+  'Антон',
+  'Аня',
+  'Толя',
+  'Коля',
+  'Женя',
+  'Ира',
+  'Катя',
+  'Соня',
+  'Зоя'
+];
+const DESCRIPTIONS = [
+  'Описание 1',
+  'Описание 2',
+  'Описание 3',
+  'Описание 4',
+  'Описание 5',
+  'Описание 6',
+  'Описание 7',
+  'Описание 8',
+  'Описание 9',
+  'Описание 10',
+  'Описание 11',
+  'Описание 12',
+  'Описание 13',
+  'Описание 14',
+  'Описание 15',
+  'Описание 16',
+  'Описание 17',
+  'Описание 18',
+  'Описание 19',
+  'Описание 20',
+  'Описание 21',
+  'Описание 22',
+  'Описание 23',
+  'Описание 24',
+  'Описание 25'
+];
+
+export{photos, photosUrl, photosId, commentId, MESSAGES, NAMES, DESCRIPTIONS};
